Delete useless code in `prover`.
